Revert "LPS-70805 Deal with integer and double numbers"
This reverts commit bb607d78e1adb2d2f043966a2fd1c9413d083863.

+++ modules/apps/forms-and-workflow/dynamic-data-mapping/dynamic-data-mapping-expression/src/main/java/com/liferay/dynamic/data/mapping/expression/internal/DDMExpressionImpl.java
@@ -153,12 +153,12 @@ public class DDMExpressionImpl<T> implements DDMExpression<T> {
 	public void setIntegerVariableValue(
 		String variableName, Integer variableValue) {
 
-		setVariableValue(variableName, variableValue.intValue());
+		setVariableValue(variableName, variableValue.doubleValue());
 	}
 
 	@Override
 	public void setLongVariableValue(String variableName, Long variableValue) {
-		setVariableValue(variableName, variableValue.longValue());
+		setVariableValue(variableName, variableValue.doubleValue());
 	}
 
 	/**
